+++ src/deal.js
@@ -1,16 +1,20 @@
 const Debug = require('debug');
 const { defaultAbiCoder, keccak256 } = require('ethers').utils;
 const {
-  isBytes32,
-  isEthAddress,
   cleanRPC,
   bnifyNestedEthersBn,
   throwIfMissing,
   http,
-  checksummedAddress,
-  ensureString,
   NULL_ADDRESS,
 } = require('./utils');
+const {
+  chainIdSchema,
+  addressSchema,
+  bytes32Schema,
+  uint256Schema,
+  positiveIntSchema,
+  positiveStrictIntSchema,
+} = require('./validator');
 
 const debug = Debug('iexec:deal');
 
@@ -22,23 +26,27 @@ const fetchRequesterDeals = async (
   } = {},
 ) => {
   try {
-    isEthAddress(requesterAddress, { strict: true });
-    if (appAddress) isEthAddress(appAddress, { strict: true });
-    if (datasetAddress) isEthAddress(datasetAddress, { strict: true });
-    if (workerpoolAddress) isEthAddress(workerpoolAddress, { strict: true });
+    const vRequesterAddress = await addressSchema().validate(requesterAddress);
+    const vChainId = await chainIdSchema().validate(chainId);
+    let vAppAddress;
+    let vDatasetAddress;
+    let vWorkerpoolAddress;
+    if (appAddress) vAppAddress = await addressSchema().validate(appAddress);
+    if (datasetAddress) vDatasetAddress = await addressSchema().validate(datasetAddress);
+    if (workerpoolAddress) vWorkerpoolAddress = await addressSchema().validate(workerpoolAddress);
     const find = Object.assign(
-      { requester: checksummedAddress(requesterAddress) },
-      appAddress && { 'app.pointer': checksummedAddress(appAddress) },
+      { requester: vRequesterAddress },
+      appAddress && { 'app.pointer': vAppAddress },
       datasetAddress && {
-        'dataset.pointer': checksummedAddress(datasetAddress),
+        'dataset.pointer': vDatasetAddress,
       },
       workerpoolAddress && {
-        'workerpool.pointer': checksummedAddress(workerpoolAddress),
+        'workerpool.pointer': vWorkerpoolAddress,
       },
       beforeTimestamp && { blockTimestamp: { $lt: beforeTimestamp } },
     );
     const body = {
-      chainId: ensureString(chainId),
+      chainId: vChainId,
       sort: {
         blockTimestamp: -1,
       },
@@ -58,14 +66,16 @@ const fetchRequesterDeals = async (
   }
 };
 
-const computeTaskId = (
+const computeTaskId = async (
   dealid = throwIfMissing(),
   taskIdx = throwIfMissing(),
 ) => {
   try {
-    isBytes32(dealid, { strict: true });
     const encodedTypes = ['bytes32', 'uint256'];
-    const values = [dealid, taskIdx];
+    const values = [
+      await bytes32Schema().validate(dealid),
+      await uint256Schema().validate(taskIdx),
+    ];
     const encoded = defaultAbiCoder.encode(encodedTypes, values);
     const taskid = keccak256(encoded);
     return taskid;
@@ -75,13 +85,18 @@ const computeTaskId = (
   }
 };
 
-const computeTaskIdsArray = (
+const computeTaskIdsArray = async (
   dealid = throwIfMissing(),
   firstTaskIdx = throwIfMissing(),
   botSize = throwIfMissing(),
 ) => {
-  const tasksIdx = [...Array(botSize).keys()].map(n => n + firstTaskIdx);
-  const taskids = tasksIdx.map(idx => computeTaskId(dealid, idx));
+  const vDealid = await bytes32Schema().validate(dealid);
+  const vFirstTaskIdx = await positiveIntSchema().validate(firstTaskIdx);
+  const vBotSize = await positiveStrictIntSchema().validate(botSize);
+  const tasksIdx = [...Array(vBotSize).keys()].map(n => n + vFirstTaskIdx);
+  const taskids = await Promise.all(
+    tasksIdx.map(idx => computeTaskId(vDealid, idx)),
+  );
   return taskids;
 };
 
@@ -90,16 +105,16 @@ const show = async (
   dealid = throwIfMissing(),
 ) => {
   try {
-    isBytes32(dealid, { strict: true });
+    const vDealid = await bytes32Schema().validate(dealid);
     const { chainId } = contracts;
     const clerkAddress = await contracts.fetchClerkAddress();
     const clerkContract = contracts.getClerkContract({ at: clerkAddress });
     const deal = bnifyNestedEthersBn(
-      cleanRPC(await clerkContract.viewDeal(dealid)),
+      cleanRPC(await clerkContract.viewDeal(vDealid)),
     );
     const dealExists = deal && deal.app && deal.app.pointer && deal.app.pointer !== NULL_ADDRESS;
     if (!dealExists) throw Error(`No deal found for dealid ${dealid} on chain ${chainId}`);
-    const tasks = computeTaskIdsArray(
+    const tasks = await computeTaskIdsArray(
       dealid,
       deal.botFirst.toString(),
       deal.botSize.toString(),
@@ -127,8 +142,6 @@ const show = async (
 
 module.exports = {
   show,
-  // claim,
   computeTaskId,
-  // computeTaskIdsArray,
   fetchRequesterDeals,
 };